requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-redis==5.2.0
djangorestframework==3.14.0
idna==3.4
kavenegar==1.1.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7
redis==4.4.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
